Use gmtime instead of own implementation.
